pyproject.toml
[project]
name = "medster"
version = "0.1.0"
description = "Autonomous Clinical Case Analysis Agent"
readme = "README.md"
requires-python = ">=3.10"
license = "MIT"
authors = [
    { name = "Your Name", email = "your.email@example.com" }
]
keywords = ["medical", "clinical", "agent", "ai", "langchain", "healthcare"]

dependencies = [
    "langchain>=0.3.0",
    "langchain-ollama>=0.1.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "prompt-toolkit>=3.0.0",
    "requests>=2.31.0",
    "pydicom>=3.0.1",
    "pillow>=12.0.0",
    "numpy>=1.24.0",
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "websockets>=12.0",
    "python-multipart>=0.0.6",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[project.scripts]
medster-agent = "medster.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/medster"]

[tool.hatch.build.targets.sdist]
include = ["src/medster"]

[tool.hatch.metadata]
allow-direct-references = true

[tool.black]
line-length = 100
target-version = ["py310"]

[tool.ruff]
line-length = 100
select = ["E", "F", "I", "N", "W"]
ignore = ["E501"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
